project/mlops-project/requirements.txt
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
mlflow==2.14.3
pyyaml==6.0.1
pytest==8.2.1
ruff==0.6.3
